% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{municipality}
\alias{municipality}
\title{Municipality data with keys and polygon-geoms for municipalities of Denmark}
\format{
A data frame with 39,230 rows and 7 columns:
 \describe{
   \item{long}{Longitude coordinates.}
   \item{lat}{Latitude coordinates.}
   \item{order}{Order of coordinates in geoms.}
   \item{group}{Geom groups.}
   \item{id}{Name of entity.}
   \item{id_numeric}{Number of entity.}
   \item{hole}{Indication of a geom hole.}
 }
}
\source{
Statistics Denmark
}
\usage{
municipality
}
\description{
Municipality data with keys and polygon-geoms for municipalities of Denmark
}
\keyword{datasets}
