% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{municipality_info}
\alias{municipality_info}
\title{Information of Valid Municipality Names and Numbers}
\format{
A data frame with 99 rows and 2 columns:
 \describe{
   \item{municipality_names}{Valid names of municipalities.}
   \item{municipality_numbers}{Valid numbers of municipalities.}
 }
}
\source{
Statistics Denmark
}
\usage{
municipality_info
}
\description{
Information of Valid Municipality Names and Numbers
}
\keyword{datasets}
