% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{province}
\alias{province}
\title{Province data with keys and polygon-geoms for provinces of Denmark}
\format{
A data frame with 4,083 rows og 7 columns:
 \describe{
   \item{long}{Longitude coordinates.}
   \item{lat}{Latitude coordinates.}
   \item{order}{Order of coordinates in geoms.}
   \item{group}{Geom groups.}
   \item{id}{Name of entity.}
   \item{id_numeric}{Number of entity.}
   \item{hole}{Indication of a geom hole.}
 }
 @source Statistics Denmark
}
\usage{
province
}
\description{
Province data with keys and polygon-geoms for provinces of Denmark
}
\keyword{datasets}
