% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{province_info}
\alias{province_info}
\title{Information of Valid Province Names and Numbers}
\format{
A data frame with 11 rows and 2 columns:
 \describe{
   \item{province_names}{Valid names of provinces.}
   \item{province_numbers}{Valid numbers of provinces.}
 }
}
\source{
Statistics Denmark
}
\usage{
province_info
}
\description{
Information of Valid Province Names and Numbers
}
\keyword{datasets}
