% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{region_info}
\alias{region_info}
\title{Information of Valid Region Names and Numbers}
\format{
A data frame with 5 observations and 2 columns:
 \describe{
   \item{region_names}{Valid names of regions.}
   \item{region_numbers}{Valid numbers of regions.}
 }
}
\source{
Statistics Denmark
}
\usage{
region_info
}
\description{
Information of Valid Region Names and Numbers
}
\keyword{datasets}
