% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zipcode_info}
\alias{zipcode_info}
\title{Information of Valid Zipcodes}
\format{
A data frame with 598 rows and 1 column:
 \describe{
   \item{zipcode_numbers}{Valid numbers of zipcodes.}
 }
}
\source{
Statistics Denmark
}
\usage{
zipcode_info
}
\description{
Information of Valid Zipcodes
}
\keyword{datasets}
