% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zipcodes}
\alias{zipcodes}
\title{Zipcode data with keys and polygon-geoms for zipcodes of Denmark}
\format{
A data frame with 49,322 rows and 7 columns:
 \describe{
   \item{long}{Longitude coordinates.}
   \item{lat}{Latitude coordinates.}
   \item{order}{Order of coordinates in geoms.}
   \item{group}{Geom groups.}
   \item{id}{Name of entity.}
   \item{id_numeric}{Number of entity.}
   \item{hole}{Indication of a geom hole.}
 }
}
\source{
Statistics Denmark
}
\usage{
zipcodes
}
\description{
Zipcode data with keys and polygon-geoms for zipcodes of Denmark
}
\keyword{datasets}
